requirements.txt
pip==23.2.1
psycopg2-binary==2.9.5
gunicorn==20.1.0
whitenoise==6.5.0

Django==3.2.21
djangorestframework == 3.14.0
python-dotenv == 0.21.1

asgiref==3.7.2
bleach==6.0.0
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
webencodings==0.5.1